refactor : types
