+++ src/pages/index.tsx
@@ -4,9 +4,9 @@ import '@scss/global.scss';
 import '@scss/pages/IndexPage.scss';
 import { useGetNotionQuery } from '@services/use-notion';
 import { getNodeJsonByUrl } from '@utils/notion';
-import { INotionContext, NotionChildrenType } from '@types';
+import { NotionChildrenType } from '@types';
 import { Contents, LatestPost } from '@components/post';
-import { NotionContext } from '@store/rootStore';
+import { INotionContext, NotionContext } from '@store/rootStore';
 import SEO from '@components/header/SEO';
 import { GlobalPortal } from '@components/GlobalPortal';
 import { StaticImage } from 'gatsby-plugin-image';
