+++ src/pages/intro.tsx
@@ -3,8 +3,8 @@ import type { HeadFC, PageProps } from 'gatsby';
 import '@scss/global.scss';
 import { useGetNotionQuery } from '@services/use-notion';
 import { getNodeJsonByUrl } from '@utils/notion';
-import { NotionChildrenType, INotionContext } from '@types';
-import { NotionContext } from '@store/rootStore';
+import { NotionChildrenType } from '@types';
+import { INotionContext, NotionContext } from '@store/rootStore';
 import SEO from '@components/header/SEO';
 import { GlobalPortal } from '@components/GlobalPortal';
 import { MainLayout } from '@layout/main';
