+++ src/pages/list.tsx
@@ -4,8 +4,8 @@ import { HeadFC, PageProps, graphql } from 'gatsby';
 import '@scss/global.scss';
 import '@scss/pages/PostsPage.scss';
 import { getNotionNodeAll } from '@services/use-notion';
-import { NotionContext } from '@store/rootStore';
-import { INotionContext, NotionNode } from '@types';
+import { INotionContext, NotionContext } from '@store/rootStore';
+import { NotionNode } from '@types';
 import { classifyPost, getParseListByNodes } from '@utils/notion';
 import SEO from '@components/header/SEO';
 import { GlobalPortal } from '@components/GlobalPortal';
