+++ src/store/rootStore.ts
@@ -1,6 +1,11 @@
 import React from 'react';
-import { INotionContext } from '@types';
+import { MultiSelect, NotionNode } from '@types';
 
+export interface INotionContext {
+  nodes: NotionNode[];
+  everyPostsTags?: string[];
+  everyPostsSeries?: MultiSelect;
+}
 export const NotionContext = React.createContext<INotionContext>({
   nodes: [],
   everyPostsTags: [],
