+++ src/types/index.ts
@@ -1,5 +1,1 @@
-export * from './nodeTypes';
-export * from './contentType';
-export * from './notionType';
-export * from './componentType';
-export * from './storeType';
+export * from './notion';
