--- src/types/storeType.ts
@@ -1,8 +1,0 @@
-import { MultiSelect } from '@types';
-import { NotionNode } from './nodeTypes';
-
-export interface INotionContext {
-  nodes: NotionNode[];
-  everyPostsTags?: string[];
-  everyPostsSeries?: MultiSelect;
-}
